Improve builder performance
